Bump configgin from 0.18.7 to 0.18.8
This pulls in a change to the code to make pods restart when the
property they pull in from a BOSH link changes.  It now correctly causes
restarts when the link they pull from is not named the same as the job
that provides that link.

See also: https://github.com/cloudfoundry-incubator/configgin/pull/104